contact page styling in the wrong branch
